LPD-15281 Transform SelectedUserHelper into a Util class, rename it to SelectedUserUtil and apply to usages
