+++ modules/apps/user-associated-data/user-associated-data-web/src/main/java/com/liferay/user/associated/data/web/internal/portlet/action/ReviewUADDataMVCRenderCommand.java
@@ -17,6 +17,7 @@ import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.user.associated.data.anonymizer.UADAnonymousUserProvider;
 import com.liferay.user.associated.data.constants.UserAssociatedDataPortletKeys;
 import com.liferay.user.associated.data.display.UADDisplay;
 import com.liferay.user.associated.data.web.internal.constants.UADConstants;
@@ -28,10 +29,10 @@ import com.liferay.user.associated.data.web.internal.display.UADEntity;
 import com.liferay.user.associated.data.web.internal.display.UADHierarchyDisplay;
 import com.liferay.user.associated.data.web.internal.display.UADInfoPanelDisplay;
 import com.liferay.user.associated.data.web.internal.display.ViewUADEntitiesDisplay;
-import com.liferay.user.associated.data.web.internal.helper.SelectedUserHelper;
 import com.liferay.user.associated.data.web.internal.helper.UADApplicationSummaryHelper;
 import com.liferay.user.associated.data.web.internal.registry.UADRegistry;
 import com.liferay.user.associated.data.web.internal.util.GroupUtil;
+import com.liferay.user.associated.data.web.internal.util.SelectedUserUtil;
 import com.liferay.user.associated.data.web.internal.util.UADSearchContainerBuilderUtil;
 
 import java.util.ArrayList;
@@ -63,7 +64,8 @@ public class ReviewUADDataMVCRenderCommand implements MVCRenderCommand {
 		throws PortletException {
 
 		try {
-			User user = _userHelper.getSelectedUser(renderRequest);
+			User user = SelectedUserUtil.getSelectedUser(
+				renderRequest, _uadAnonymousUserProvider);
 
 			List<ScopeDisplay> scopeDisplays = _getScopeDisplays(user);
 
@@ -316,12 +318,12 @@ public class ReviewUADDataMVCRenderCommand implements MVCRenderCommand {
 	private Portal _portal;
 
 	@Reference
+	private UADAnonymousUserProvider _uadAnonymousUserProvider;
+
+	@Reference
 	private UADApplicationSummaryHelper _uadApplicationSummaryHelper;
 
 	@Reference
 	private UADRegistry _uadRegistry;
-
-	@Reference
-	private SelectedUserHelper _userHelper;
 
 }
